+++ Assets/Scripts/Difficulty.cs
@@ -7,8 +7,8 @@ public class Difficulty
     readonly int id;
     readonly string nom;
     readonly int roomLength;
-    readonly double dmgPlayer;
-    readonly double dmgMob;
+    readonly decimal dmgPlayer;
+    readonly decimal dmgMob;
     readonly int spawnMob;
     readonly int minSpawnWall;
     readonly int maxSpawnWall;
@@ -19,8 +19,8 @@ public class Difficulty
     public int Id { get => id; }
     public string Nom { get => nom; }
     public int RoomLength { get => roomLength; }
-    public double DmgPlayer { get => dmgPlayer; }
-    public double DmgMob { get => dmgMob; }
+    public decimal DmgPlayer { get => dmgPlayer; }
+    public decimal DmgMob { get => dmgMob; }
     public int SpawnMob { get => spawnMob; }
     public int MinSpawnWall { get => minSpawnWall; }
     public int MaxSpawnWall { get => maxSpawnWall; }
@@ -33,7 +33,7 @@ public class Difficulty
     static public Difficulty selected;
     static public List<Difficulty> difficulties;
 
-    public Difficulty(int Id, string Nom, int RoomLength, double DmgPlayer, double DmgMob, int SpawnMob, int MinSpawnWall, int MaxSpawnWall, int MinSpawnFood, int MaxSpawnFood, decimal CoefMoney)
+    public Difficulty(int Id, string Nom, int RoomLength, decimal DmgPlayer, decimal DmgMob, int SpawnMob, int MinSpawnWall, int MaxSpawnWall, int MinSpawnFood, int MaxSpawnFood, decimal CoefMoney)
     {
         id = Id;
         nom = Nom;
